requirements.txt
Django>=1.8,<2.0
django-allauth==0.30.0
django-rest-auth==0.9.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-extensions==1.7.5
psycopg2
django-webpack-loader